requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
langchain-core>=0.3.0
streamlit>=1.38.0
requests>=2.31.0
numpy>=1.26.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
prometheus-client>=0.21.0
flask>=3.0.0
